# docs/swagger.yaml
basePath: /
definitions:
...
  dto.CreateCard:
    properties:
      description:
        type: string
      position:
        type: string
      skills:
        items:
          type: string
        type: array
    required:
    - description
    - position
    - skills
    type: object
...
  dto.UpdateCard:
    properties:
      description:
        type: string
      position:
        type: string
      skills:
        items:
          type: string
        type: array
    required:
    - description
    - position
    - skills
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Backend for AITU Corporate Self-Study Portal.
  title: ACSP Backend
  version: "1.0"
paths:
...
  /api/v1/code-connection/applicants:
    get:
      consumes:
      - application/json
      description: Get all cards
      operationId: get-all-cards
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
        default:
          description: ""
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Get all cards
      tags:
      - cards
  /api/v1/code-connection/applicants/{id}:
    get:
      consumes:
      - application/json
      description: Get applicant card by id
      operationId: get-applicant-card-by-id
      parameters:
      - description: applicant id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
        default:
          description: ""
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Get applicant card by id
      tags:
      - cards
  /api/v1/code-connection/cards:
    get:
      consumes:
      - application/json
      description: Get all cards of a user
      operationId: get-all-cards-by-user-id
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
        default:
          description: ""
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Get all cards of a user
      tags:
      - cards
    post:
      consumes:
      - application/json
      description: Method for creating a card
      operationId: card-id
      parameters:
      - description: card information
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateCard'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: integer
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
        default:
          description: ""
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Create a card
      tags:
      - cards
  /api/v1/code-connection/cards/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a card from database
      operationId: delete-card-by-id
      parameters:
      - description: Card ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
        default:
          description: ""
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Delete a card by id
      tags:
      - cards
    put:
      consumes:
      - application/json
      description: Update data of card
      operationId: update-card-by-id
      parameters:
      - description: card id
        in: path
        name: id
        required: true
        type: integer
      - description: card information
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateCard'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
        default:
          description: ""
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Update a card by id
      tags:
      - cards
  /api/v1/code-connection/cards/{id}/invitations:
    get:
      consumes:
      - application/json
      description: Get all invitations by card id
      operationId: get-invitations-by-card-id
      parameters:
      - description: card id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
        default:
          description: ""
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Get all invitations by card
      tags:
      - cards
    post:
      consumes:
      - application/json
      description: Invite a user by a card id
      operationId: create-an-invitation
      parameters:
      - description: card id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            type: integer
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
        default:
          description: ""
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Create an invitation
      tags:
      - cards
  /api/v1/code-connection/cards/{id}/invitations/{invitationID}:
    get:
      consumes:
      - application/json
      description: Get invitation by id and card id
      operationId: get-invitation-by-id
      parameters:
      - description: card id
        in: path
        name: id
        required: true
        type: integer
      - description: invitation id
        in: path
        name: invitationID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
        default:
          description: ""
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Get invitation by id and card id
      tags:
      - cards
  /api/v1/code-connection/cards/{id}/invitations/{invitationID}/accept:
    post:
      consumes:
      - application/json
      description: Accept invitation by invitation id and card id
      operationId: accept-invitation-by-id-and-card-id
      parameters:
      - description: card id
        in: path
        name: id
        required: true
        type: integer
      - description: invitation id
        in: path
        name: invitationID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
        default:
          description: ""
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Accept invitation by invitation id and card id
      tags:
      - cards
  /api/v1/code-connection/cards/{id}/invitations/{invitationID}/decline:
    post:
      consumes:
      - application/json
      description: Decline invitation by invitation id and card id
      operationId: decline-invitation-by-id-and-card-id
      parameters:
      - description: card id
        in: path
        name: id
        required: true
        type: integer
      - description: invitation id
        in: path
        name: invitationID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
        default:
          description: ""
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Decline invitation by invitation id and card id
      tags:
      - cards
  /api/v1/code-connection/cards/invitations:
    get:
      consumes:
      - application/json
      description: Get invitations of a user's cards
      operationId: get-card-invitations
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
        default:
          description: ""
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Get invitations of user
      tags:
      - cards
  /api/v1/code-connection/cards/responses:
    get:
      consumes:
      - application/json
      description: Get all responses of a user
      operationId: get-responses-by-user-id
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
        default:
          description: ""
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Get responses of a user
      tags:
      - cards
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"